+++ index.html
@@ -12,7 +12,6 @@
 		<link rel="icon" type="image/png" href="/images/favicons/favicon-512.png" sizes="512x512" />
 		<link rel="shortcut icon" href="/images/favicons/favicon.ico">
 		<link rel="apple-touch-icon" href="/images/favicon-512.png" />
-		<link rel="stylesheet" href="/src/ui/index.css" />
 	</head>
 	<body>
