+++ jsconfig.json
@@ -2,5 +2,6 @@
 	"compilerOptions": {
 		"checkJs": true,
 		"target": "esnext",
+		"moduleResolution": "nodenext"
 	},
 }
