+++ src/main.js
@@ -1,7 +1,1 @@
-import {Component, render} from 'preact'
-import {useState} from 'preact/hooks'
-import {html} from 'htm/preact'
-
-// This file bundles the things we need in the UI to make the imports easier.
-
-export {Component, render, useState, html}
+import './ui/styles/index.css'
