--- src/ui/index.css
@@ -1,20 +1,0 @@
-@import './styles/app.css';
-@import './styles/fonts.css';
-@import './styles/overlay.css';
-@import './styles/card.css';
-@import './styles/targets.css';
-@import './styles/map.css';
-@import './styles/tooltipped.css';
-
-:root {
-	--card-width: 11em;
-	--card-height: 17em;
-	--yellow: #ffc107;
-	--teal: #116f54;
-	--lightblue: hsl(194.7, 53.3%, 79%);
-	--red: hsl(0, 77%, 29%);
-	--bg: hsl(0, 30%, 8%);
-	--text: hsl(34, 78%, 85%);
-	--text-inverse: hsla(0, 0%, 0%, 0.95);
-	--text-stroke: var(--text-inverse);
-}
