+++ src/ui/overlays.js
@@ -1,4 +1,4 @@
-import {html} from '../main.js'
+import {html} from './lib.js'
 
 // A tiny overlay UI component.
 export function Overlay(props) {
