+++ src/ui/splash-screen.js
@@ -1,4 +1,4 @@
-import {html, Component} from '../main.js'
+import {html, Component} from './lib.js'
 import gsap from './animations.js'
 
 export default class SplashScreen extends Component {
